Extract const-related funcs into module
