+++ lib/skc_async_experiment/src/hir_to_mir.rs
@@ -1,8 +1,8 @@
 mod collect_allocs;
+mod constants;
 use crate::names::FunctionName;
 use crate::{build, hir, mir};
 use anyhow::Result;
-use shiika_core::names::ConstFullname;
 use shiika_core::ty::TermTy;
 use skc_hir::SkTypes;
 use std::collections::HashSet;
@@ -29,9 +29,13 @@ pub fn run(
         .map(convert_method)
         .collect();
     log::debug!("User functions converted");
-    funcs.push(create_const_inits(
+    funcs.push(constants::create_const_inits(
         hir.package_name.as_ref(),
-        hir.program.constants,
+        hir.program
+            .constants
+            .into_iter()
+            .map(|(name, rhs)| (name, convert_texpr(rhs)))
+            .collect(),
     ));
     if let build::CompileTargetDetail::Bin { total_deps, .. } = &target.detail {
         funcs.push(create_user_main(hir.program.top_exprs, total_deps));
@@ -219,38 +223,12 @@ fn convert_expr(expr: hir::Expr<TermTy>) -> mir::Expr {
     }
 }
 
-fn create_const_inits(
-    package: Option<&String>,
-    constants: Vec<(ConstFullname, hir::TypedExpr<TermTy>)>,
-) -> mir::Function {
-    let mut body_stmts: Vec<_> = constants
-        .into_iter()
-        .map(|(name, rhs)| mir::Expr::const_set(mir::mir_const_name(name), convert_texpr(rhs)))
-        .collect();
-    body_stmts.push(mir::Expr::return_(mir::Expr::number(0)));
-    mir::Function {
-        asyncness: mir::Asyncness::Unknown,
-        name: package_const_init_name(package),
-        params: vec![],
-        ret_ty: mir::Ty::Raw("Int".to_string()),
-        body_stmts: mir::Expr::exprs(body_stmts),
-    }
-}
-
 fn create_user_main(
     top_exprs: Vec<hir::TypedExpr<TermTy>>,
     total_deps: &[String],
 ) -> mir::Function {
     let mut body_stmts = vec![];
-    body_stmts.extend(total_deps.iter().map(|name| {
-        let fname = package_const_init_name(Some(name));
-        let fun_ty = mir::FunTy::new(
-            mir::Asyncness::Unknown,
-            vec![],
-            mir::Ty::Raw("Int".to_string()),
-        );
-        mir::Expr::fun_call(mir::Expr::func_ref(fname, fun_ty), vec![])
-    }));
+    body_stmts.extend(constants::call_all_const_inits(total_deps));
     body_stmts.extend(top_exprs.into_iter().map(convert_texpr));
     body_stmts.push(mir::Expr::return_(mir::Expr::number(0)));
     mir::Function {
@@ -260,13 +238,4 @@ fn create_user_main(
         ret_ty: mir::Ty::Raw("Int".to_string()),
         body_stmts: mir::Expr::exprs(body_stmts),
     }
-}
-
-fn package_const_init_name(package_name: Option<&String>) -> FunctionName {
-    let suffix = if let Some(pkg) = package_name {
-        format!("_{}", pkg)
-    } else {
-        String::new()
-    };
-    FunctionName::mangled(format!("shiika_init_const_{}", suffix))
 }
